% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{rainfallMap}
\alias{rainfallMap}
\title{A randomly generated field representing a continuous rainfall gradient}
\format{A two-dimensional matrix with 100 rows and 100 columns
\describe{
  ...
}}
\usage{
rainfallMap
}
\description{
A randomly generated field representing a continuous rainfall gradient
}
\keyword{datasets}
