% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{temperatureMap}
\alias{temperatureMap}
\title{A randomly generated field representing a continuous temperature gradient}
\format{A two-dimensional matrix with 100 rows and 100 columns
\describe{
  ...
}}
\usage{
temperatureMap
}
\description{
A randomly generated field representing a continuous temperature gradient
}
\keyword{datasets}
